Cleanup: extract UV alignment into its own file
UpdateBrushFaceAttributes.h/.cpp held two unrelated concerns: the update command
vocabulary and the UV alignment geometry that merely returns such an update.
Only four of the 21 files that included the header needed the latter.

Move the UvPolicy, UvFitMode, UvAxis and UvSign enums, anchorVertex, isAligned,
isJustified, isFitted, align, justify and fit, along with their private helpers,
into UvAlignment.h/.cpp, and split the tests accordingly. The dependency is
one-directional: UvAlignment includes UpdateBrushFaceAttributes, not the
reverse.

This is a pure move; no logic changes.
